Add `nix-watch` lock file (#8)
Solves a critical problem in which the mod_time variable was being
overlooked. Uses a lock file to check against for update times instead,
which stores the path that changed and the modification time. Also adds
a graceful shutdown which removes the PID_FILE, and removes the
LOCK_FILE if DEBUG is false.

Tested on aarch64-darwin and x86_64-linux, both are working as expected.

Closes #4
Closes #7